Add Proxy Featur Fix missing fav Icon

+++ scripts/entrypoint.sh
@@ -2,6 +2,12 @@
 
 echo $1
 
+if [[ ! -f $1 ]]
+then
+    echo "$1 does not exist on your filesystem. EXITING"
+    exit 1
+fi
+
 cp $1 ./app/backend/dataset.json
 
 echo "Starting Backend  ..."
